processPlaceQueryForm method in PlaceController refactored

+++ src/main/java/pl/warkoczewski/foodspot/dto/DisplayPlaceDTO.java
@@ -3,13 +3,11 @@ package pl.warkoczewski.foodspot.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @AllArgsConstructor
 @Data
 public class DisplayPlaceDTO {
-    private BigDecimal lon;
-    private BigDecimal lat;
+    private Double lon;
+    private Double lat;
     private String name;
     private String address;
