+++ src/main/java/pl/warkoczewski/foodspot/service/validation/ValidationService.java
@@ -3,6 +3,6 @@ package pl.warkoczewski.foodspot.service.validation;
 import java.math.BigDecimal;
 
 public interface ValidationService {
-    boolean isLatitudeBetween(BigDecimal latitude);
-    boolean isLongitudeBetween(BigDecimal longitude);
+    boolean isLatitudeBetween(Double latitude);
+    boolean isLongitudeBetween(Double longitude);
 }
