+++ src/main/java/pl/warkoczewski/foodspot/util/Coordinates.java
@@ -3,10 +3,10 @@ package pl.warkoczewski.foodspot.util;
 import java.math.BigDecimal;
 
 public final class Coordinates {
-    public static final BigDecimal LATITUDE_MIN = BigDecimal.valueOf(-90);
-    public static final BigDecimal LATITUDE_MAX = BigDecimal.valueOf(90);
-    public static final BigDecimal LONGITUDE_MIN = BigDecimal.valueOf(-180);
-    public static final BigDecimal LONGITUDE_MAX = BigDecimal.valueOf(180);
+    public static final Double LATITUDE_MIN = -90d;
+    public static final Double LATITUDE_MAX = 90d;
+    public static final Double LONGITUDE_MIN = -180d;
+    public static final Double LONGITUDE_MAX = 180d;
 
     public Coordinates() {
     }
