+++ src/main/java/pl/warkoczewski/foodspot/validation/validator/LongitudeValidatorForBigDecimal.java
@@ -10,7 +10,7 @@ import javax.validation.ConstraintValidatorContext;
 import java.math.BigDecimal;
 @Component
 @Scope("prototype")
-public class LongitudeValidatorForBigDecimal implements ConstraintValidator<Longitude, BigDecimal> {
+public class LongitudeValidatorForBigDecimal implements ConstraintValidator<Longitude, Double> {
     private final ValidationServiceImpl validationService;
 
     public LongitudeValidatorForBigDecimal(ValidationServiceImpl validationService) {
@@ -20,7 +20,7 @@ public class LongitudeValidatorForBigDecimal implements ConstraintValidator<Long
     public void initialize(Longitude constraintAnnotation) {
     }
     @Override
-    public boolean isValid(BigDecimal longitude, ConstraintValidatorContext context) {
+    public boolean isValid(Double longitude, ConstraintValidatorContext context) {
         return validationService.isLongitudeBetween(longitude);
     }
 }
